seperate the components of herosection
