+++ src/app/globals.css
@@ -3,102 +3,3 @@
 
 /* Tailwind v4 import */
 @import "tailwindcss";
-
-/* Tailwind v4 theme configuration */
-@theme {
-  /* Typography */
-  --font-family-sans: "Inter", ui-sans-serif, system-ui, sans-serif;
-  --font-family-serif: "Crimson Text", Georgia, serif;
-  --font-family-mono: "JetBrains Mono", Monaco, Consolas, monospace;
-
-  /* Colors using OKLCH (v4 preferred format) */
-  --color-background: oklch(100% 0 0); /* Pure white */
-  --color-foreground: oklch(15% 0.02 220); /* Dark text */
-  
-  --color-card: oklch(100% 0 0); /* White cards */
-  --color-card-foreground: oklch(15% 0.02 220);
-  
-  --color-primary: oklch(56% 0.18 220); /* Blue */
-  --color-primary-foreground: oklch(100% 0 0); /* White text on blue */
-  --color-primary-glow: oklch(70% 0.15 220);
-  
-  --color-secondary: oklch(55% 0.15 140); /* Green */
-  --color-secondary-foreground: oklch(100% 0 0);
-  
-  --color-muted: oklch(96% 0.01 210); /* Light gray */
-  --color-muted-foreground: oklch(45% 0.02 220); /* Medium gray */
-  
-  --color-accent: oklch(95% 0.02 200); /* Light blue */
-  --color-accent-foreground: oklch(25% 0.03 220); /* Dark blue */
-  
-  --color-border: oklch(90% 0.01 220); /* Light border */
-  --color-input: oklch(95% 0.01 220); /* Input background */
-  --color-ring: oklch(55% 0.15 220); /* Focus ring */
-  
-  --color-destructive: oklch(60% 0.17 0); /* Red */
-  --color-destructive-foreground: oklch(100% 0 0);
-  
-  --color-popover: oklch(100% 0 0); /* White popover */
-  --color-popover-foreground: oklch(15% 0.02 220);
-
-  /* Chart colors */
-  --color-chart-1: oklch(55% 0.15 220);
-  --color-chart-2: oklch(65% 0.14 200);
-  --color-chart-3: oklch(55% 0.15 140);
-  --color-chart-4: oklch(65% 0.15 260);
-  --color-chart-5: oklch(55% 0.15 30);
-
-  /* Radius */
-  --radius: 0.75rem;
-  --radius-sm: 0.5rem;
-  --radius-md: 0.625rem;
-  --radius-lg: 0.75rem;
-  --radius-xl: 1rem;
-}
-
-/* Dark mode theme (if needed) */
-@media (prefers-color-scheme: dark) {
-  @theme {
-    --color-background: oklch(5% 0.02 222);
-    --color-foreground: oklch(98% 0.01 210);
-    --color-card: oklch(5% 0.02 222);
-    --color-card-foreground: oklch(98% 0.01 210);
-  }
-}
-
-/* Base styles */
-* {
-  box-sizing: border-box;
-}
-
-body {
-  font-family: theme(fontFamily.sans);
-  background: white;
-  color: rgb(38 38 38);
-  line-height: 1.5;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  opacity: 0;
-  animation: fadeIn 300ms ease-out forwards;
-}
-
-html {
-  scroll-behavior: smooth;
-  scroll-padding-top: 5rem;
-}
-
-/* Simple animations */
-@keyframes fadeIn {
-  to { opacity: 1; }
-}
-
-@keyframes float {
-  0%, 100% { transform: translateY(0) rotate(0deg); }
-  33% { transform: translateY(-10px) rotate(1deg); }
-  66% { transform: translateY(-5px) rotate(-1deg); }
-}
-
-/* KaTeX styling */
-.question-text .katex { font-size: 1.15em !important; }
-.question-text-modern .katex { font-size: 1.1em !important; }
-.question-text-mono .katex { font-size: 1.05em !important; }
